requirements.txt
# requirements.txt (example with versions)
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.1.0
pymongo==4.5.0
python-dotenv==1.0.0
pytesseract==0.3.10
pyzbar==0.1.9
Werkzeug==2.3.7
python-doctr[torch]
